src/requirements.txt
amqp==5.1.1
anyio==3.6.2
APScheduler==3.6.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
autopep8==2.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
boto3==1.26.115
botocore==1.29.115
cachetools==4.2.2
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==7.3.0
cryptography==40.0.2
defusedxml==0.7.1
Django==4.1
django-cors-headers==4.2.0
django-debug-toolbar==4.0.0
django-environ==0.10.0
django-extensions==3.2.1
django-filter==23.2
django-silk==5.0.3
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.3
execnet==2.0.2
Faker==19.1.0
flake8==6.1.0
gevent==23.7.0
gprof2dot==2022.7.29
greenlet==2.0.2
gunicorn==20.0.4
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
iniconfig==2.0.0
install==1.3.5
jmespath==1.0.1
kombu==5.2.4
lxml==4.9.3
mccabe==0.7.0
oauthlib==3.2.2
packaging==23.2
Pillow==9.5.0
pluggy==1.3.0
prompt-toolkit==3.0.38
psycopg2-binary==2.8.6
pycodestyle==2.11.0
pycparser==2.21
pyflakes==3.1.0
PyJWT==2.6.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-django==4.6.0
pytest-xdist==3.3.1
python-dateutil==2.8.2
python-telegram-bot==13.6
python3-openid==3.2.0
pytz==2023.3
requests==2.30.0
s3transfer==0.6.0
sentry-sdk==1.29.2
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
tomli==2.0.1
tornado==6.3.2
typing_extensions==4.7.1
tzlocal==5.0.1
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.4.0
zope.event==5.0
zope.interface==6.0
